pyproject.toml
[project]
name = "mycrew"
version = "0.1.0"
description = "mycrew - AI-powered development team using CrewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[litellm,tools]==1.9.3",
    "pyyaml>=6.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0,<1.0.0",
    "pydantic-settings>=2.10.1,<2.11",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
kickoff = "mycrew.main:kickoff"
run_crew = "mycrew.main:kickoff"
plot = "mycrew.main:plot"
run_with_trigger = "mycrew.main:run_with_trigger"
webhook = "mycrew.webhook:main"
register_webhook = "mycrew.register_webhook:main"
kickoff-client = "mycrew.kickoff_client:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/code_pipeline"]
omit = [
    "src/code_pipeline/main.py",  # Flow orchestration; tested via integration
]

[tool.coverage.report]
fail_under = 75

[tool.ruff.lint.per-file-ignores]
"src/code_pipeline/main.py" = ["E402"]

[tool.uv]
# onnxruntime 1.24+ dropped macOS x86_64 wheels; 1.19.2 has universal2 (Intel + ARM)
override-dependencies = ["onnxruntime==1.19.2"]
